+++ src/builtins/mod.rs
@@ -1,0 +1,5 @@
+pub mod calc;
+pub mod cd;
+pub mod echo;
+pub mod help;
+pub mod ls;
